requirements.txt
asgiref==3.11.1
Django==5.2.1
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
cryptography==46.0.7
sqlparse==0.5.5
tzdata==2026.1
google-genai>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv>=1.0.0
cffi==2.0.0
pycparser==3.0
